pyproject.toml
[project]
name = "opencode-platform"
version = "1.0.0"
description = "OpenCode-Centric Intelligent Platform - 企業級自主代理系統"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OpenCode Team"}
]
keywords = ["ai", "agent", "rag", "mcp", "llm", "orchestrator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Async
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    
    # CLI/TUI
    "typer>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    
    # AI/LLM
    "openai>=1.0.0",
    
    # Vector DB
    "qdrant-client>=1.7.0",
    
    # Document Processing
    "docling>=2.0.0",
    
    # Storage
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    
    # Web Framework
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "python-multipart>=0.0.6",
    
    # MCP
    "mcp>=1.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
docker = [
    "docker>=6.1.0",
]

[project.scripts]
opencode = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "core", "orchestrator", "gateway", "services", "control_plane", "api", "config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
